+++ __tests__/bundle/bundle-lint-format/json-format-snapshot.js
@@ -7,7 +7,7 @@ exports[`E2E bundle lint format bundle lint: should be formatted by format: json
     "warnings": 0,
     "ignored": 0
   },
-  "version": "1.0.0-beta.64",
+  "version": "1.0.0-beta.65",
   "problems": [
     {
       "ruleId": "spec",
@@ -31,7 +31,7 @@ exports[`E2E bundle lint format bundle lint: should be formatted by format: json
     "warnings": 0,
     "ignored": 0
   },
-  "version": "1.0.0-beta.64",
+  "version": "1.0.0-beta.65",
   "problems": []
 }
 ❌ Validation failed with 1 error.